fix: Apply automatic clippy fixes across project
- Fix unnecessary casting in Ch10/generics (42 as f64 → 42_f64)
- Fix redundant slicing in Ch10/lifetimes (use original value instead of &[..])
- Fix needless lifetimes in Ch10/lifetimes (elide explicit lifetimes)
- Fix let_and_return pattern in Ch10/lifetimes
- Fix useless vec! usage in multiple files (use arrays instead)
- Fix empty line after doc comments in Ch12/testing
- Fix double_ended_iterator_last in Ch12/accepting_arguments
- Fix items_after_test_module in Ch12/environment_variables
- Fix match_like_matches_macro in Ch12/environment_variables
- Remove unused imports in Ch12/reading_files and Ch9/unrecoverable_errors
- Fix get_first usage in Ch9/exercises and Ch9/unrecoverable_errors
- Fix manual_abs_diff in method_syntax examples
- Fix iter_nth_zero in Ch9/error_propagation

Total: 94 insertions, 107 deletions across 14 files

+++ rust_book/Ch10/generics/src/main.rs
@@ -271,7 +271,7 @@ fn example7_user_input_and_generics() {
     // process_number(42, 42.0);  // Compile error!
     
     // ✅ But you CAN convert between types:
-    let converted_float = process_number(42 as f64);
+    let converted_float = process_number(42_f64);
     println!("Converted and processed: {}", converted_float);
 
     println!();
